+++ pytket/extensions/qiskit/backends/ibm.py
@@ -51,7 +51,7 @@ from pytket.passes import (
     AutoRebase,
     BasePass,
     CliffordSimp,
-    CXMappingPass,
+    CustomPass,
     DecomposeBoxes,
     FullPeepholeOptimise,
     KAKDecomposition,
@@ -61,7 +61,6 @@ from pytket.passes import (
     SimplifyInitial,
     SynthesiseTket,
 )
-from pytket.placement import NoiseAwarePlacement
 from pytket.predicates import (
     DirectednessPredicate,
     GateSetPredicate,
@@ -94,7 +93,7 @@ from ..qiskit_convert import (
     tk_to_qiskit,
 )
 from .config import QiskitConfig
-from .ibm_utils import _STATUS_MAP, _batch_circuits
+from .ibm_utils import _STATUS_MAP, _batch_circuits, _gen_lightsabre_transformation
 
 if TYPE_CHECKING:
     from qiskit_ibm_runtime.ibm_backend import IBMBackend  # type: ignore
@@ -345,7 +344,6 @@ class IBMQBackend(Backend):
     def default_compilation_pass(
         self,
         optimisation_level: int = 2,
-        placement_options: Optional[dict[str, Any]] = None,
     ) -> BasePass:
         """
         A suggested compilation pass that will will, if possible, produce an equivalent
@@ -359,8 +357,6 @@ class IBMQBackend(Backend):
         is tailored to the backend's requirements.
 
         The default compilation passes for the :py:class:`IBMQBackend` and the
-        Aer simulators support an optional ``placement_options`` dictionary containing
-        arguments to override the default settings in :py:class:`NoiseAwarePlacement`.
 
         :param optimisation_level: The level of optimisation to perform during
             compilation.
@@ -372,14 +368,12 @@ class IBMQBackend(Backend):
               that should give the best results from execution.
 
 
-        :param placement_options: Optional argument allowing the user to override
-          the default settings in :py:class:`NoiseAwarePlacement`.
         :return: Compilation pass guaranteeing required predicates.
         """
         config: PulseBackendConfiguration = self._backend.configuration()
         props: Optional[BackendProperties] = self._backend.properties()
         return IBMQBackend.default_compilation_pass_offline(
-            config, props, optimisation_level, placement_options
+            config, props, optimisation_level
         )
 
     @staticmethod
@@ -387,7 +381,6 @@ class IBMQBackend(Backend):
         config: PulseBackendConfiguration,
         props: Optional[BackendProperties],
         optimisation_level: int = 2,
-        placement_options: Optional[dict[str, Any]] = None,
     ) -> BasePass:
         backend_info = IBMQBackend._get_backend_info(config, props)
         primitive_gates = _get_primitive_gates(_tk_gate_set(config))
@@ -409,33 +402,12 @@ class IBMQBackend(Backend):
             passlist.append(SynthesiseTket())
         elif optimisation_level == 2:
             passlist.append(FullPeepholeOptimise())
-        mid_measure = backend_info.supports_midcircuit_measurement
         arch = backend_info.architecture
         assert arch is not None
         if not isinstance(arch, FullyConnected):
-            if placement_options is not None:
-                noise_aware_placement = NoiseAwarePlacement(
-                    arch,
-                    backend_info.averaged_node_gate_errors,  # type: ignore
-                    backend_info.averaged_edge_gate_errors,  # type: ignore
-                    backend_info.averaged_readout_errors,  # type: ignore
-                    **placement_options,
-                )
-            else:
-                noise_aware_placement = NoiseAwarePlacement(
-                    arch,
-                    backend_info.averaged_node_gate_errors,  # type: ignore
-                    backend_info.averaged_edge_gate_errors,  # type: ignore
-                    backend_info.averaged_readout_errors,  # type: ignore
-                )
-
+            passlist.append(AutoRebase(primitive_gates))
             passlist.append(
-                CXMappingPass(
-                    arch,
-                    noise_aware_placement,
-                    directed_cx=True,
-                    delay_measures=(not mid_measure),
-                )
+                CustomPass(_gen_lightsabre_transformation(arch, optimisation_level))
             )
             passlist.append(NaivePlacementPass(arch))
         if optimisation_level == 1:
@@ -449,11 +421,10 @@ class IBMQBackend(Backend):
                 ]
             )
 
-        if supports_rz:
-            passlist.extend(
-                [IBMQBackend.rebase_pass_offline(primitive_gates), RemoveRedundancies()]
-            )
-        return SequencePass(passlist)
+        passlist.extend(
+            [IBMQBackend.rebase_pass_offline(primitive_gates), RemoveRedundancies()]
+        )
+        return SequencePass(passlist, False)
 
     @property
     def _result_id_type(self) -> _ResultIdTuple: