+++ pytket/extensions/qiskit/backends/ibmq_emulator.py
@@ -15,7 +15,6 @@
 from collections.abc import Sequence
 from typing import (
     TYPE_CHECKING,
-    Any,
     Optional,
 )
 
@@ -88,13 +87,12 @@ class IBMQEmulatorBackend(Backend):
     def default_compilation_pass(
         self,
         optimisation_level: int = 2,
-        placement_options: Optional[dict[str, Any]] = None,
     ) -> BasePass:
         """
         See documentation for :py:meth:`IBMQBackend.default_compilation_pass`.
         """
         return self._ibmq.default_compilation_pass(
-            optimisation_level=optimisation_level, placement_options=placement_options
+            optimisation_level=optimisation_level
         )
 
     @property